third_party/WebKit/LayoutTests/editing/pasteboard/file-drag-to-editable-expected.txt

Tests that dragging a file into an editable area does not insert a filename.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.attemptedFileNavigation is true
PASS successfullyParsed is true

TEST COMPLETE
